+++ lib/src/format_number_util.dart
@@ -15,22 +15,22 @@
 import 'package:decimal/decimal.dart';
 
 part 'formatter_config.dart';
-part 'decimal_ext.dart';
+part 'format_number_ext.dart';
 
 String formatNumber(
   Decimal? val, {
-  int precision = 0, // 展示精度
+  int precision = 0, // 精度
   bool showSign = false, // 是否展示符号位+
   RoundMode? roundMode,
   bool cutInvalidZero = false, // 删除尾部零
   bool enableCompact = false, // 是否启用精简转换器转换大数展示, 优先于千分位展示
   CompactConverter? compactConverter, // 自定义整数部分的精简转换器
-  bool enableGrouping = false, // 是否启用分组, 主要是针对千分位展示.
-  String? groupSepartor, // 定制分组分隔符, 默认','.
-  int? groupCounts, // 定制分组数量, 默认3.
-  ShrinkZeroMode? shrinkZeroMode, // 小数部分多零情况下, 进行收缩展示模式.
-  ShrinkZeroConverter? shrinkZeroConverter, // 自定义多零收缩转换器.
-  UnicodeDirection? direction, // 如果不为空, 则对结果进行ltr或rtl
+  bool enableGrouping = false, // 是否启用分组, 主要是针对千分位展示
+  String? groupSepartor, // 定制分组分隔符, 默认','
+  int? groupCounts, // 定制分组数量, 默认3
+  ShrinkZeroMode? shrinkZeroMode, // 小数部分多零情况下, 进行收缩展示模式
+  ShrinkZeroConverter? shrinkZeroConverter, // 自定义多零收缩转换器
+  ExplicitDirection? direction, // 如果不为空, 则对结果应用显式双向格式
   String prefix = '', // 前缀
   String suffix = '', // 后缀
   String? defIfZero, // 如果为0时的默认展示
@@ -38,19 +38,19 @@ String formatNumber(
 }) {
   // 处理数据为空的情况
   if (val == null) {
-    return '$prefix$defIfNull$suffix'.unicodeDirection(direction);
+    return '$prefix$defIfNull$suffix'.applyExplicitBidiFormatting(direction);
   }
 
   // 处理数据为0的情况
   if (val == Decimal.zero && defIfZero != null) {
-    return '$prefix$defIfZero$suffix'.unicodeDirection(direction);
+    return '$prefix$defIfZero$suffix'.applyExplicitBidiFormatting(direction);
   }
 
   if (showSign && val >= Decimal.zero) prefix += '+';
 
   String ret;
   if (enableCompact) {
-    final (result, unit) = val._compact(
+    final (result, unit) = val.compact(
       precision: precision,
       roundMode: roundMode,
       isClean: cutInvalidZero,
@@ -62,7 +62,7 @@ String formatNumber(
     );
     suffix = unit + suffix;
   } else if (enableGrouping) {
-    var (integerPart, decimalPart) = val._group(
+    var (integerPart, decimalPart) = val.group(
       precision,
       roundMode: roundMode,
       isClean: cutInvalidZero,
@@ -86,7 +86,7 @@ String formatNumber(
     );
   }
 
-  return '$prefix$ret$suffix'.unicodeDirection(direction);
+  return '$prefix$ret$suffix'.applyExplicitBidiFormatting(direction);
 }
 
 /// 百分比
@@ -97,15 +97,15 @@ String formatPercentage(
   bool showSign = false,
   RoundMode roundMode = RoundMode.floor,
   bool cutInvalidZero = false,
-  UnicodeDirection? direction,
+  ExplicitDirection? direction,
   String suffix = '%',
   String defIfNull = '--', // 如果为空或无效值时的默认展示.
 }) {
   return formatNumber(
     val == null ? null : (expandHundred ? val * hundred : val),
     precision: precision,
-    roundMode: roundMode,
     showSign: showSign,
+    roundMode: roundMode,
     cutInvalidZero: cutInvalidZero,
     direction: direction,
     suffix: suffix,
@@ -123,7 +123,7 @@ String formatPrice(
   bool enableGrouping = true,
   ShrinkZeroMode? shrinkZeroMode,
   ShrinkZeroConverter? shrinkZeroConverter,
-  UnicodeDirection? direction,
+  ExplicitDirection? direction,
   String prefix = '',
   String suffix = '',
   String? defIfZero,
@@ -157,7 +157,7 @@ String formatAmount(
   bool cutInvalidZero = true,
   ShrinkZeroMode? shrinkZeroMode,
   ShrinkZeroConverter? shrinkZeroConverter,
-  UnicodeDirection? direction,
+  ExplicitDirection? direction,
   String prefix = '',
   String suffix = '',
   String? defIfZero,